Add files via upload
Receive commands via serial communication to control a stepper motor connected to an Arduino. The commands can be "1" to rotate the motor clockwise, "2" to rotate counterclockwise, and "0" to stop the motor. Depending on the command received, the program activates the motor in the desired direction or stops it. The motor rotates a certain number of steps in the specified direction and then stops for a period of time, allowing you to control the speed and direction of rotation of the motor.